Tower Selection
I have added logic to select a tower. That code is in the Player.cs
Update method. For this, some fields have been added in the tower class
and in the player class and in the sprite class as well.

Next thing to be done is to display the information of the selected
tower on the toolbar.